CLOUD-50403 make ConstraintTemplate consistent with other resources
