+++ core-api/src/main/java/com/sequenceiq/cloudbreak/api/model/ConstraintTemplateResponse.java
@@ -5,10 +5,16 @@ import com.sequenceiq.cloudbreak.doc.ModelDescriptions;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
+import javax.validation.constraints.NotNull;
+
 @ApiModel
 public class ConstraintTemplateResponse extends ConstraintTemplateBase {
     @ApiModelProperty(ModelDescriptions.ID)
     private Long id;
+
+    @NotNull
+    @ApiModelProperty(value = ModelDescriptions.PUBLIC_IN_ACCOUNT, readOnly = true)
+    private boolean publicInAccount;
 
     public Long getId() {
         return id;
@@ -16,5 +22,13 @@ public class ConstraintTemplateResponse extends ConstraintTemplateBase {
 
     public void setId(Long id) {
         this.id = id;
+    }
+
+    public boolean isPublicInAccount() {
+        return publicInAccount;
+    }
+
+    public void setPublicInAccount(boolean publicInAccount) {
+        this.publicInAccount = publicInAccount;
     }
 }
